backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
sqlmodel==0.0.16
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
python-jose==3.3.0
fastapi-admin==1.0.3
tortoise-orm==0.20.0
aioredis==2.0.1
jinja2==3.0.3 